+++ X10D.Tests/src/Math/UInt16Tests.cs
@@ -32,6 +32,28 @@ public class UInt16Tests
     }
 
     [TestMethod]
+    public void GreatestCommonFactor_ShouldBe1_ForPrimeNumbers()
+    {
+        const ushort first = 5;
+        const ushort second = 7;
+
+        ushort multiple = first.GreatestCommonFactor(second);
+
+        Assert.AreEqual(1, multiple);
+    }
+
+    [TestMethod]
+    public void GreatestCommonFactor_ShouldBe6_Given12And18()
+    {
+        const ushort first = 12;
+        const ushort second = 18;
+
+        ushort multiple = first.GreatestCommonFactor(second);
+
+        Assert.AreEqual(6, multiple);
+    }
+
+    [TestMethod]
     public void IsEvenShouldBeCorrect()
     {
         const ushort one = 1;