+++ X10D/src/Math/SByteExtensions.cs
@@ -64,6 +64,23 @@ public static class SByteExtensions
     }
 
     /// <summary>
+    ///     Calculates the greatest common factor between the current 8-bit signed integer, and another 8-bit signed integer.
+    /// </summary>
+    /// <param name="value">The first value.</param>
+    /// <param name="other">The second value.</param>
+    /// <returns>The greatest common factor between <paramref name="value" /> and <paramref name="other" />.</returns>
+    [Pure]
+#if NETSTANDARD2_1
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#else
+    [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+#endif
+    public static sbyte GreatestCommonFactor(this sbyte value, sbyte other)
+    {
+        return (sbyte)((long)value).GreatestCommonFactor(other);
+    }
+
+    /// <summary>
     ///     Returns a value indicating whether the current value is evenly divisible by 2.
     /// </summary>
     /// <param name="value">The value whose parity to check.</param>